requirements.txt
fastapi>=0.115
uvicorn>=0.30
pydantic>=2.8
numpy>=1.26
sentence-transformers>=3.0
datapizza-ai>=0.0.7
datapizza-ai-clients-openai>=0.0.3
python-dotenv>=1.0

